+++ js/data.js
@@ -1,0 +1,56 @@
+const DESCR = [
+  'А я на море',
+  'Я и моя тень',
+  'Солнечный день',
+  'C любовью',
+  'Год спустя',
+  'Плохая погода',
+  'Экзамен',
+  'Вашингтон',
+];
+
+const AUTHOR = [
+  'Markus',
+  'Anton',
+  'Igor',
+  'Nastya',
+  'Sveta',
+  'Inokentii',
+  'Yana',
+  'Olya',
+  'Irina',
+  'Konstantin'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const PHOTO_COUNT = 25;
+
+const Likes = {
+  MIN: 15,
+  MAX: 135
+};
+
+const Comments = {
+  MIN: 1,
+  MAX: 2
+};
+
+const Id = {
+  MIN: 1,
+  MAX: 25
+};
+
+const Pic = {
+  MIN: 1,
+  MAX: 6
+};
+
+export {DESCR, AUTHOR, MESSAGE, PHOTO_COUNT, Likes, Comments, Id, Pic};
